test: 인증 플로우 E2E 테스트 구현 및 테스트 구조 개선
## 변경 사항

### 1. 테스트 디렉토리 구조 개선
- playwright/ → tests/로 테스트 파일 이동
- playwright.config.ts의 testDir 경로 업데이트
- 테스트 파일 구조를 더 명확하게 정리

### 2. 인증 플로우 완전 테스트 구현
테스트 계획서(frontend/specs/basic-user-workflows-test-plan.md)의
"1. 인증 플로우" 섹션의 모든 시나리오를 Playwright로 구현

#### 구현된 테스트 시나리오 (5개)
- 1.1 회원가입 - 정상 케이스
  * 고유한 이메일로 회원가입
  * 자동 로그인 및 워크스페이스 생성 확인

- 1.2 회원가입 - 중복 이메일
  * 중복 이메일 오류 처리 확인
  * 회원가입 페이지 유지 검증

- 1.3 로그인 - 정상 케이스
  * 로그인 성공 및 리다이렉트 확인
  * JWT 토큰(accessToken) 저장 검증

- 1.4 로그인 - 잘못된 자격증명
  * 로그인 실패 처리 확인
  * 로그인 페이지 유지 검증

- 1.5 로그아웃
  * 로그아웃 성공 및 리다이렉트 확인
  * 토큰 제거 및 사이드바 숨김 검증

### 3. 테스트 안정성 개선
- 백엔드 health check를 /api/auth/login으로 변경
  (Spring Actuator 엔드포인트 의존성 제거)
- localStorage 토큰 키를 'accessToken'으로 정확히 수정
- 테스트 격리를 위한 타임스탬프 기반 고유 이메일 생성
- 각 테스트 시나리오에 상세한 한국어 주석 추가

## 테스트 결과
- 총 15개 테스트 (Chromium 5개, Firefox 5개, WebKit 5개)
- ✅ 100% 통과 (15/15)
- 실행 시간: 16.5초

## 관련 파일
- frontend/tests/auth-flow.spec.ts (새 파일)
- frontend/tests/helpers.ts (playwright/에서 이동)
- frontend/playwright.config.ts (testDir 수정)
- frontend/specs/basic-user-workflows-test-plan.md (이전 커밋)
